+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Updated to storefront api v2
+### Added
+- New configuration: saveCustomerData
+
 ## [1.1.4] - 2022-04-12
 
 ### Fix: